Product gallery update feature

+++ app/Http/Livewire/Admin/AdminProductComponent.php
@@ -19,11 +19,13 @@ class AdminProductComponent extends Component
 
         if ($product->images) {
             $images = explode(',', $product->images);
+
             foreach ($images as $image) {
-                unlink('assets/images/products' . '/' . $image);
+                if ($image) {
+                    unlink('assets/images/products' . '/' . $image);
+                }
             }
         }
-
         $product->delete();
 
         session()->flash('message', 'Product deleted successfully!');
